pyproject.toml
[project]
name = "dusrabheja"
version = "2.0.0"
description = "Discord-based personal AI command center / second brain"
requires-python = ">=3.12"
dependencies = [
    # Discord
    "discord.py>=2.4",

    # Web framework
    "fastapi>=0.115",
    "uvicorn>=0.34",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "alembic>=1.14",

    # Job queue
    "arq>=0.26",

    # AI
    "anthropic>=0.42",
    "openai>=1.58",

    # MCP
    "mcp>=1.0",

    # File processing
    "pymupdf4llm>=0.0.17",
    "openpyxl>=3.1",

    # Config
    "pydantic-settings>=2.7",

    # HTTP
    "httpx>=0.28",

    # File I/O
    "aiofiles>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
